refactor: when node dead, persister sql migrate other node
