+++ query_server/query/Cargo.toml
@@ -53,6 +53,7 @@ strum_macros = { workspace = true }
 tempfile = { workspace = true }
 tokio = { workspace = true, features = ["full", "tracing"] }
 tokio-util = { workspace = true }
+tokio-retry = { workspace = true }
 url = { workspace = true }
 
 [features]
