+++ query_server/query/src/dispatcher/mod.rs
@@ -1,6 +1,6 @@
 use async_trait::async_trait;
-use spi::query::dispatcher::QueryInfo;
-use spi::service::protocol::QueryId;
+use models::meta_data::NodeId;
+use models::schema::query_info::{QueryId, QueryInfo};
 use spi::QueryResult;
 
 pub mod manager;
@@ -11,5 +11,5 @@ pub mod query_tracker;
 pub trait QueryPersister {
     fn remove(&self, query_id: &QueryId) -> QueryResult<()>;
     async fn save(&self, query_id: QueryId, query: QueryInfo) -> QueryResult<()>;
-    async fn queries(&self) -> QueryResult<Vec<QueryInfo>>;
+    async fn queries(&self, node_id: NodeId) -> QueryResult<Vec<QueryInfo>>;
 }
