+++ query_server/query/src/execution/query.rs
@@ -2,8 +2,9 @@ use std::sync::Arc;
 
 use async_trait::async_trait;
 use futures::stream::AbortHandle;
+use models::schema::query_info::QueryInfo;
 use parking_lot::Mutex;
-use spi::query::dispatcher::{QueryInfo, QueryStatus};
+use spi::query::dispatcher::QueryStatus;
 use spi::query::execution::{Output, QueryExecution, QueryStateMachineRef};
 use spi::query::logical_planner::QueryPlan;
 use spi::query::optimizer::Optimizer;
@@ -108,6 +109,7 @@ impl QueryExecution for SqlQueryExecution {
             qsm.session.tenant().to_string(),
             qsm.session.default_database().to_string(),
             qsm.session.user().clone(),
+            qsm.coord.node_id(),
         )
     }
