+++ query_server/query/src/execution/sys/mod.rs
@@ -3,7 +3,8 @@ mod kill_query;
 use std::sync::Arc;
 
 use async_trait::async_trait;
-use spi::query::dispatcher::{QueryInfo, QueryStatus};
+use models::schema::query_info::QueryInfo;
+use spi::query::dispatcher::QueryStatus;
 use spi::query::execution::{Output, QueryExecution, QueryStateMachineRef};
 use spi::query::logical_planner::SYSPlan;
 use spi::QueryResult;
@@ -64,6 +65,7 @@ impl QueryExecution for SystemExecution {
             qsm.session.tenant().to_string(),
             qsm.session.default_database().to_string(),
             qsm.session.user().clone(),
+            qsm.coord.node_id(),
         )
     }
     // 运行时信息
