+++ query_server/query/src/stream/watermark_tracker.rs
@@ -8,6 +8,7 @@ use datafusion::arrow::record_batch::RecordBatch;
 use futures::TryStreamExt;
 use models::arrow::TimeUnit;
 use models::predicate::domain::Predicate;
+use models::schema::query_info::QueryId;
 use models::schema::tskv_table_schema::{ColumnType, TableColumn, TskvTableSchema};
 use models::schema::{CLUSTER_SCHEMA, DEFAULT_CATALOG};
 use models::utils::now_timestamp_nanos;
@@ -16,7 +17,6 @@ use protocol_parser::Line;
 use protos::FieldValue;
 use snafu::ResultExt;
 use spi::query::session::SessionCtx;
-use spi::service::protocol::QueryId;
 use spi::{CoordinatorSnafu, ModelsSnafu, PersistQuerySnafu, QueryError};
 use trace::debug;
 use tskv::reader::QueryOption;
