+++ query_server/spi/src/service/protocol.rs
@@ -1,64 +1,10 @@
-use std::fmt::Display;
-
 use models::auth::user::User;
-use models::oid::uuid_u64;
+use models::schema::query_info::QueryId;
 use models::schema::{DEFAULT_CATALOG, DEFAULT_DATABASE, DEFAULT_PRECISION};
-use serde::{Deserialize, Serialize};
 
 use crate::query::config::StreamTriggerInterval;
 use crate::query::execution::Output;
 use crate::query::session::CnosSessionConfig;
-
-#[derive(Default, Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub struct QueryId(u64);
-
-impl QueryId {
-    pub fn next_id() -> Self {
-        Self(uuid_u64())
-    }
-
-    pub fn get(&self) -> u64 {
-        self.0
-    }
-}
-
-impl From<QueryId> for u64 {
-    fn from(val: QueryId) -> Self {
-        val.0
-    }
-}
-
-impl From<u64> for QueryId {
-    fn from(u: u64) -> Self {
-        QueryId(u)
-    }
-}
-
-impl TryFrom<Vec<u8>> for QueryId {
-    type Error = String;
-
-    fn try_from(bytes: Vec<u8>) -> Result<Self, Self::Error> {
-        if bytes.len() != 8 {
-            return Err(format!("Incorrect content: {:?}", &bytes));
-        }
-
-        let len_bytes: [u8; 8] = unsafe { bytes[0..8].try_into().unwrap_unchecked() };
-
-        Ok(Self(u64::from_le_bytes(len_bytes)))
-    }
-}
-
-impl From<QueryId> for Vec<u8> {
-    fn from(val: QueryId) -> Self {
-        val.0.to_le_bytes().into()
-    }
-}
-
-impl Display for QueryId {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
 
 #[derive(Clone)]
 pub struct Context {
